requirements.txt
python==3.12
requests==2.32.3
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
statsmodels==0.14.3
xgboost==2.1.1
streamlit==1.38.0
python-dotenv==1.0.1